pyproject.toml
[project]
name = "hackathon-reply"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "faiss-cpu>=1.13.2",
    "langchain>=1.2.15",
    "langchain-community>=0.4.1",
    "langchain-huggingface>=1.2.1",
    "langchain-openai>=1.1.14",
    "langfuse>=4.3.1",
    "numpy>=2.4.4",
    "pandas>=3.0.2",
    "pydantic>=2.13.1",
    "python-dotenv>=1.2.2",
    "sentence-transformers>=5.4.1",
    "ulid-py>=1.1.0",
]
